Fixed compression button bug
Originally, the compression button would set the slider to the necessary
value, which would then be taken to the text box. The issue with this,
however, is that the compression value is then subjected to being
rounded to the nearest (total range of slider)/100, in this case, 2.
This has been fixed